+++ Equilibrium/CardTemplate.cs
@@ -8,23 +8,22 @@ using UnboundLib.Cards;
 using UnityEngine;
 
 
-namespace Equilibrium.Cards
+namespace Equilibrium
 {
-    class Template : CustomCard
+    class CardTemplate : CustomCard
     {
         public override void SetupCard(CardInfo cardInfo, Gun gun, ApplyCardStats cardStats, CharacterStatModifiers statModifiers, Block block)
         {
-            //Edits values on card itself, which are then applied to the player in `ApplyCardStats`
+            
         }
         public override void OnAddCard(Player player, Gun gun, GunAmmo gunAmmo, CharacterData data, HealthHandler health, Gravity gravity, Block block, CharacterStatModifiers characterStats)
         {
-            //Edits values on player when card is selected
+            
         }
         public override void OnRemoveCard(Player player, Gun gun, GunAmmo gunAmmo, CharacterData data, HealthHandler health, Gravity gravity, Block block, CharacterStatModifiers characterStats)
         {
-            //Run when the card is removed from the player
+            
         }
-
 
         protected override string GetTitle()
         {