+++ Equilibrium/Cards/Evasion.cs
@@ -75,7 +75,7 @@ namespace Equilibrium.Cards
 
         protected override CardThemeColor.CardThemeColorType GetTheme()
         {
-            return CardThemeColor.CardThemeColorType.DefensiveBlue;
+            return CardThemeColor.CardThemeColorType.TechWhite;
         }
         public override string GetModName()
         {