+++ Equilibrium/Cards/Tag.cs
@@ -79,7 +79,7 @@ namespace Equilibrium.Cards
         }
         protected override CardThemeColor.CardThemeColorType GetTheme()
         {
-            return CardThemeColor.CardThemeColorType.EvilPurple;
+            return CardThemeColor.CardThemeColorType.MagicPink;
         }
         public override string GetModName()
         {