+++ Equilibrium/Component/HealthStackMono.cs
@@ -1,0 +1,64 @@
+﻿using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Text;
+using UnityEngine;
+
+namespace Equilibrium.Component
+{
+    public class HealthStackMono : MonoBehaviour
+    {
+        public CharacterData data;
+        public HealthHandler health;
+
+        public float baseFactor = 0.25f;
+        public float increment = 0f;
+        public float currentMultiplier;
+
+        public bool resetPerRound = true;
+
+        private float lastHealth;
+
+        void Start()
+        {
+            data = GetComponent<CharacterData>();
+            StartCoroutine(WaitForHealth());
+        }
+
+        private IEnumerator WaitForHealth()
+        {
+            while (data.healthHandler == null)
+                yield return null;
+
+            health = data.healthHandler;
+        }
+
+        void Update()
+        {
+            if (health == null) return;
+            if (data.health != lastHealth)
+            {
+                if (data.health < lastHealth)
+                {
+                    OnDamageTaken(lastHealth - data.health);
+                }
+                lastHealth = data.health;
+            }
+        }
+
+        private void OnDamageTaken(float damage)
+        {
+            data.maxHealth += baseFactor * increment * damage;
+        }
+
+        public void AddIncrement()
+        {
+            increment += 1f;
+        }
+
+        void OnDestroy()
+        {
+            increment = 0f;
+        }
+    }
+}